refactor: restructure services package
