+++ chat-service/src/main/java/com/example/chat_service/services/chat/ChatService.java
@@ -1,4 +1,4 @@
-package com.example.chat_service.services;
+package com.example.chat_service.services.chat;
 
 import com.example.chat_service.models.Chat;
 import com.example.chat_service.models.Message;
