+++ chat-service/src/main/java/com/example/chat_service/services/observer/MessageSubject.java
@@ -1,4 +1,4 @@
-package com.example.chat_service.services;
+package com.example.chat_service.services.observer;
 
 import com.example.chat_service.models.Message;
